ImageCache Thread Safety
Made ImageCache thread safe
Switched to returning shared_ptrs instead of references in order to be able to keep track of if something is still using an image from a cache. This is to support potentially having multiple caches and multiple processors running in parallel all running at different rates.