ImageCacheProvider
Added ImageCacheProvider which is a container of in-use ImageCaches.
Creates a new cache whenever it is reset with an image. As long as someone is using an image from a cache (shared_ptr use_count), the cache will continue to exist within ImageCacheProvider. Old caches are cleaned periodically.
The idea is there will be one ImageCacheProvider that will somehow be updated with new image from the camera. All image processors will go through the ICP to get an image to process.