VIC-13524 VIC-14584 VIC-13523 CompositeImage refactor (#5547)
* VIC-13524 Merge SpriteSeqTrack into SpriteBoxCompositor

These changes pass responsibility for the last few things handled by
the older CompositeImage system in the Anim process on to the
SpriteBoxCompositor. Specifically, handling legacy FaceAnimKeyFrames
and the SetFaceImage functionality.

Now no longer in use, the CompositeImage infrastructure is removed
from the anim process, data loaders, etc. wherever possible, leaving
it in place only as a last stop for rendering images to the face.

* VIC-14584 Fix lingering spriteSequence anims

If an anim was played using the _proceduralAnim in the
AnimationStreamer, it was possible for it to leave some
SpriteBoxKeyFrames in place if they were not cleared during run time.
The next feature to use the _proceduralAnim to display something, e.g
using the SetFaceImage function, would then end up with residual
SpriteBoxKeyFrames from the last animation if it didn't first clear
the animation, which is not always possible. This change simply clears
whatever is left in the _proceduralAnim when it finishes playing,
since the case where it is aborted(interrupted) is already covered.

* VIC-13523 refactor CompImage around SpriteBoxCompositor model

Strips down the CompositeImage class to limit its concerns to:
1. Providing a container of linked SpriteBoxes and SpriteHandles
representing a group of images to be rendered to the screen, and all
the information to render them, for one static "frame"
2. Providing the logic for how those images should be rendered given
their settings and layers

After refactoring, these changes also clean up dependencies and remove
structures, classes and CLAD definitions which are no longer necessary.