VIC-13273: Downgrade Robot.UpdateFullRobotState.GetLastPoseWithFrameIdError (#5557)
It's not surprising that this error would sometimes appear on startup given that the history is cleared on start. Suppressing when frame_id == 0 eliminates the bulk of the errors. There are a few instances where frame_id doesn't equal 0, but they seem transient and they don't appear to be causing problems. Downgrading to warning. If we start seeing them in high volume then more investigation is required.